Refactor Image in multiple choice question
