+++ WebApp/frontend/quiz_session/app/room/quiz/components/modal.js
@@ -1,0 +1,22 @@
+import React from "react";
+import { Modal } from "@mantine/core";
+import Image from "next/image";
+
+export default function ImageModal({ opened, close, imageUrl }) {
+  return (
+    <Modal
+      opened={opened}
+      onClose={close}
+      withCloseButton={false}
+      centered
+      size="xl"
+    >
+      <Image
+        src={imageUrl}
+        width={0}
+        height={0}
+        style={{ width: "100%", height: "100%" }}
+      />
+    </Modal>
+  );
+}
