+++ WebApp/frontend/quiz_session/app/util/api.js
@@ -11,3 +11,29 @@ export const submitAnswer = ({ id, answer, connectionId }) => {
     }),
   });
 };
+
+export const downloadImage = async ({ url, setImageUrl }) => {
+  try {
+    const authToken = localStorage.getItem("token");
+
+    const response = await fetch(`${url}`, {
+      headers: {
+        Authorization: `Bearer ${authToken}`,
+      },
+    });
+
+    // Check if the request was successful (status code 200)
+    if (response.ok) {
+      const blob = await response.blob();
+      const imageUrl = URL.createObjectURL(blob);
+
+      // Set the image URL in the state or use it directly in the component
+      setImageUrl(imageUrl);
+    } else {
+      // Handle errors
+      console.error("Error downloading image:", response.statusText);
+    }
+  } catch (error) {
+    console.error("Error downloading image:", error);
+  }
+};
